small ads refactor
