+++ src/index.js
@@ -1,5 +1,5 @@
 import mux from 'mux-embed';
-
+import initializeKalturaImaEvents from './ads.js';
 const log = mux.log;
 const secondsToMs = mux.utils.secondsToMs;
 const assign = mux.utils.assign;
@@ -17,7 +17,6 @@ const initKalturaMux = function (player, options) {
   }
 
   const PlaybackEventMap = new Map();
-  const AdsEventMap = new Map();
 
   PlaybackEventMap.set('play', player.Event.Core.PLAY);
   PlaybackEventMap.set('videochange', player.Event.Core.CHANGE_SOURCE_STARTED);
@@ -28,17 +27,6 @@ const initKalturaMux = function (player, options) {
   PlaybackEventMap.set('seeked', player.Event.Core.SEEKED);
   PlaybackEventMap.set('ended', player.Event.Core.ENDED);
   PlaybackEventMap.set('error', player.Event.Core.ERROR);
-
-  AdsEventMap.set('adresponse', player.Event.AD_LOADED);
-  AdsEventMap.set('adbreakstart', player.Event.AD_BREAK_START);
-  AdsEventMap.set('adplaying', player.Event.AD_STARTED);
-  AdsEventMap.set('adpause', player.Event.AD_PAUSED);
-  AdsEventMap.set('adfirstquartile', player.Event.AD_FIRST_QUARTILE);
-  AdsEventMap.set('admidpoint', player.Event.AD_MIDPOINT);
-  AdsEventMap.set('adthirdquartile', player.Event.AD_THIRD_QUARTILE);
-  AdsEventMap.set('adended', player.Event.AD_COMPLETED);
-  AdsEventMap.set('adbreakend', player.Event.AD_BREAK_END);
-  AdsEventMap.set('aderror', player.Event.AD_ERROR);
 
   // Prepare the data passed in
   options = options || {};
@@ -56,7 +44,7 @@ const initKalturaMux = function (player, options) {
   // Enable customers to emit events through the player instance
   player.mux = {};
   player.mux.emit = function (eventType, data) {
-   // mux.emit(playerID, eventType, data);
+    // mux.emit(playerID, eventType, data);
     console.log('EMIT:', playerID, eventType, data);
   };
 
@@ -105,25 +93,10 @@ const initKalturaMux = function (player, options) {
     });
   });
 
-  AdsEventMap.forEach((kalturaEvent, muxEvent) => {
-    player.addEventListener(kalturaEvent, (event) => {
-      let data = {};
-
-      if (kalturaEvent === player.Event.AD_LOADED) {
-        const ad_tag_url = player.plugins.ima.config.adTagUrl;
-
-        data.ad_tag_url = ad_tag_url;
-      } if (kalturaEvent === player.Event.AD_STARTED) {
-        const ad_asset_url = player.ads.getAd()._url;
-
-        data.ad_asset_url = ad_asset_url;
-      } if (kalturaEvent === player.Event.AD_ERROR) {
-        data.player_error_code = event.payload.code;
-        data.player_error_message = event.payload.data.message;
-      }
-      player.mux.emit(muxEvent, data);
-    });
-  });
+	// Register ad plugins
+  if (player.plugins.ima) {
+    initializeKalturaImaEvents(player);
+  }
 
   // Initialize the tracking
   // mux.init(playerID, options);
